refactor: publicly expose two impl "with_http_req" and "with_sp_offchain"
Needed for "pallet-ocw-circuits" and "pallet-ocw-garble".
Completely split "with_http_req" vs "with_sp_offchain".
This simplifies the logic; and gives full control to the callers (ie in repo "pallets")

- rename create "http-grpc-client" -> "http-client"
- add feature "with_sp_offchain"
- move implementations to separate mods
- rename some structs
- rename crate: http_grpc_client -> interstellar_http_client
- ci:
    - remove protoc
    - use a matrix
- ![no_std]: ipfs-client-http-req
