+++ ipfs-client-http-req/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "ipfs-client-http-req"
-version = "0.2.0"
+name = "interstellar-ipfs-client"
+version = "0.5.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -9,18 +9,30 @@ edition = "2021"
 log = { version = "0.4", default-features = false }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-serde_with = { version = "2.1.0", default-features = false, features = ["alloc", "macros"] }
+serde_with = { version = "2.1.0", default-features = false, features = [
+    "alloc",
+    "macros",
+] }
 snafu = { version = "0.7", default-features = false }
 # http_req_sgx = { package = "http_req", git = "https://github.com/mesalock-linux/http_req-sgx", optional = true }
 # default is "native-tls", but that needs to (cross-)compile openssl-sys which COULD be a pain on Android; In ancy case, rustls works fine
 # http_req_std = { package = "http_req", version = "0.9", optional = true, default-features = false, features = ["rust-tls"] }
 sgx_tstd = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
-http-grpc-client = { default-features = false, git = "https://github.com/Interstellar-Network/rs-common.git", branch = "main" }
+interstellar-http-client = { default-features = false, path = "../http-client" }
 
 [features]
 default = ["std"]
-sgx = ["http-grpc-client/sgx", "sgx_tstd", "http-grpc-client/with_http_req_sgx"]
-std = ["http-grpc-client/std", "http-grpc-client/with_http_req_std"]
+sgx = [
+    "interstellar-http-client/sgx",
+    "sgx_tstd",
+    "interstellar-http-client/with_http_req_sgx",
+]
+std = [
+    "interstellar-http-client/std",
+    "interstellar-http-client/with_http_req_std",
+]
+with_http_req = ["interstellar-http-client/with_http_req"]
+with_sp_offchain = ["interstellar-http-client/with_sp_offchain"]
 
 [dev-dependencies]
 tests-utils = { git = "https://github.com/Interstellar-Network/rs-common", branch = "main" }
